+++ index.js
@@ -1,0 +1,2 @@
+export {LitNotify} from "./notify.js";
+export {bind} from "./bind.js";
